Add a SpawnerDataBuilder interface and implementation
This interface allows for easy, single-line construction of SpawnerData instances which was previously either impossible or very difficult to do in a clean way. An instance of SpawnerData was required and a sequence of methods had to be invoked in order to get the desired effect. The builder pattern should alleviate these issues.